refactor(generate-roadmap): rename /ai route to /generate-roadmap
