+++ apps/web/app/(full-layout)/generate-roadmap/_components/generation-loading.tsx
