+++ apps/web/app/(full-layout)/generate-roadmap/_types/personalized-questions-chat.types.ts
