+++ apps/web/app/(full-layout)/generate-roadmap/_validations/ai-roadmap.schema.ts
